finished api with authentication

+++ controllers/ContactController.js
@@ -3,16 +3,16 @@ const Contact = require('../models/contactModel');
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
 //@desc Create new contact
 //@route CREATE /api/contacts
-//@access public
+//@access private
 
 const createContact = asyncHandler(async (req, res) => {
   const { name, email, phone } = req.body;
@@ -21,6 +21,7 @@ const createContact = asyncHandler(async (req, res) => {
     throw new Error('All fields are mandatory!');
   }
   const contact = await Contact.create({
+    user_id: req.user.id,
     name,
     email,
     phone,
@@ -30,7 +31,7 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc Get individual contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 
 const getContact = asyncHandler(async (req, res) => {
   const contacts = await Contact.findById(req.params.id);
@@ -43,20 +44,38 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Update individual contact
 //@route UPDATE /api/contacts/:id
-//@access public
+//@access private
 
 const updateContact = asyncHandler(async (req, res) => {
   const { id } = req.params;
+  const contact = await Contact.findById(id);
+  if (!contacts) {
+    res.status(404);
+    throw new Error('Contact not found');
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error('Not Authorized');
+  }
   const contacts = await Contact.findByIdAndUpdate(id, req.body, { new: true });
   res.status(200).json(contacts);
 });
 
 //@desc Delete individual contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 
 const deleteContact = asyncHandler(async (req, res) => {
   const { id } = req.params;
+  const contact = await Contact.findById(id);
+  if (!contacts) {
+    res.status(404);
+    throw new Error('Contact not found');
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error('Not Authorized');
+  }
   const contacts = await Contact.findByIdAndDelete(id);
   if (!contacts) {
     res.status(404);
